Processing split and external validation
This release changes the processing of files so that each file is processed separately and opened independently so that the vectors are created for that file alone and the questions are made using only that file's data.

The results are then inserted into an output file in csv format that can be used for a script validation.